+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+### Added
+- generate initial README.md when none exists in a project
+
 ## [0.3.0] - 2026-03-16
 
 ### Changed